Simple rename of hex_print
